+++ src/main.rs
@@ -6,7 +6,7 @@ use thwack::{entrypoint, safe_exit, DefaultTerminal};
 fn main() {
     let mut out = stdout();
     let err = stderr();
-    match entrypoint(env::args_os(), &mut out, DefaultTerminal) {
+    match entrypoint(env::args_os(), DefaultTerminal, &mut out) {
         Ok(_) => safe_exit(0, out, err),
         Err(e) => {
             eprintln!("{}", e); // TODO: Write a more readable error message.
